Added support for oAuth for YQL lookups
oAuth is automatically enabled when the cc_key and cc_secret keys are
set to something in the config.ini config file.

When enabled oAuth mode causes class.geoservice.php to alter which end
point URL it uses, and also adds the appropriate oAuth headers to the
HTTP request.

Finally, because the rate limits on the oAuth API are much looser than
on the non-oAuth API, the webServiceWait variable is reduced from 0.5
seconds to 0.2 seconds. Note: this does have the danger of hitting the
100K/day limit, but keeps within the 20K/hour per IP limit.